+++ NEWS.md
@@ -1,3 +1,7 @@
+# ggdemetra 0.2.7
+
+- deprecated function `y_forecast()` removed.
+
 # ggdemetra 0.2.6
 
 * `siratio()` correction with TRAMO-SEATS `"jSA"` models.